// frontend/src/api/generated/types.gen.ts
export type NameCount = {
    count?: number;
    name?: string;
};